Move code into fs.go and templates.go
