+++ condor.go
@@ -30,7 +30,6 @@
 package main
 
 import (
-	"bytes"
 	"encoding/json"
 	"flag"
 	"fmt"
@@ -62,65 +61,26 @@ var log = logrus.WithFields(logrus.Fields{
 })
 
 func init() {
+	var err error
 	logrus.SetFormatter(&logrus.JSONFormatter{})
+	SubmissionTemplate, err = template.New("condor_submit").Parse(SubmissionTemplateText)
+	if err != nil {
+		log.Fatal(errors.Wrap(err, "failed to parse submission template text"))
+	}
+
+	JobConfigTemplate, err = template.New("job_config").Parse(JobConfigTemplateText)
+	if err != nil {
+		log.Fatal(errors.Wrap(err, "failed to parse job config template text"))
+	}
+
+	IRODSConfigTemplate, err = template.New("irods_config").Parse(IRODSConfigTemplateText)
+	if err != nil {
+		log.Fatal(errors.Wrap(err, "failed to parse irods config template text"))
+	}
 }
 
 const pingKey = "events.condor-launcher.ping"
 const pongKey = "events.condor-launcher.pong"
-
-// SubmissionTemplateText is the text of the template for the HTCondor
-// submission file.
-const SubmissionTemplateText = `universe = vanilla
-executable = /usr/local/bin/road-runner
-rank = mips{{ if .UsesVolumes }}
-requirements = (HAS_HOST_MOUNTS == True){{ end }}
-arguments = --config config --job job
-output = script-output.log
-error = script-error.log
-log = condor.log{{if .Group}}
-accounting_group = {{.Group}}
-accounting_group_user = {{.Submitter}}{{end}}
-request_disk = {{.RequestDisk}}
-+IpcUuid = "{{.InvocationID}}"
-+IpcJobId = "generated_script"
-+IpcUsername = "{{.Submitter}}"
-+IpcUserGroups = {{.FormatUserGroups}}
-concurrency_limits = {{.UserIDForSubmission}}
-{{with $x := index .Steps 0}}+IpcExe = "{{$x.Component.Name}}"{{end}}
-{{with $x := index .Steps 0}}+IpcExePath = "{{$x.Component.Location}}"{{end}}
-should_transfer_files = YES
-transfer_input_files = irods-config,iplant.cmd,config,job
-transfer_output_files = workingvolume/logs/logs-stdout-output,workingvolume/logs/logs-stderr-output
-when_to_transfer_output = ON_EXIT_OR_EVICT
-notification = NEVER
-queue
-`
-
-// JobConfigTemplateText is the text of the template for the HTCondor submission
-// file.
-const JobConfigTemplateText = `amqp:
-    uri: {{.GetString "amqp.uri"}}
-    exchange:
-        name: {{.GetString "amqp.exchange.name"}}
-        type: {{.GetString "amqp.exchange.type"}}
-irods:
-    base: "{{.GetString "irods.base"}}"
-porklock:
-    image: "{{.GetString "porklock.image"}}"
-    tag: "{{.GetString "porklock.tag"}}"
-condor:
-    filter_files: "{{.GetString "condor.filter_files"}}"`
-
-// IRODSConfigTemplateText is the text of the template for porklock's iRODS
-// config file.
-const IRODSConfigTemplateText = `porklock.irods-host = {{.IRODSHost}}
-porklock.irods-port = {{.IRODSPort}}
-porklock.irods-user = {{.IRODSUser}}
-porklock.irods-pass = {{.IRODSPass}}
-porklock.irods-home = {{.IRODSBase}}
-porklock.irods-zone = {{.IRODSZone}}
-porklock.irods-resc = {{.IRODSResc}}
-`
 
 // Messenger defines an interface for handling AMQP operations. This is the
 // subset of functionality needed by job-status-recorder.
@@ -133,138 +93,32 @@ type Messenger interface {
 	PublishJobUpdate(*messaging.UpdateMessage) error
 }
 
-// fsys defines an interface for file operations.
-type fsys interface {
-	MkdirAll(string, os.FileMode) error
-	WriteFile(string, []byte, os.FileMode) error
-}
-
-// osys is an implementation of fsys that hits the os and ioutil packages.
-type osys struct{}
-
-func (o *osys) MkdirAll(path string, mode os.FileMode) error {
-	return os.MkdirAll(path, mode)
-}
-
-func (o *osys) WriteFile(path string, contents []byte, mode os.FileMode) error {
-	return ioutil.WriteFile(path, contents, mode)
-}
-
 // CondorLauncher contains the condor-launcher application state.
 type CondorLauncher struct {
-	cfg                 *viper.Viper
-	client              Messenger
-	fs                  fsys
-	submissionTemplate  *template.Template
-	jobConfigTemplate   *template.Template
-	irodsConfigTemplate *template.Template
+	cfg    *viper.Viper
+	client Messenger
+	fs     fsys
 }
 
 // New returns a new *CondorLauncher
-func New(c *viper.Viper, client Messenger, fs fsys) (*CondorLauncher, error) {
-	cl := &CondorLauncher{
+func New(c *viper.Viper, client Messenger, fs fsys) *CondorLauncher {
+	return &CondorLauncher{
 		cfg:    c,
 		client: client,
 		fs:     fs,
 	}
-	st, err := template.New("condor_submit").Parse(SubmissionTemplateText)
-	if err != nil {
-		return cl, errors.Wrap(err, "failed to parse submission template text")
-	}
-	cl.submissionTemplate = st
-	jct, err := template.New("job_config").Parse(JobConfigTemplateText)
-	if err != nil {
-		return cl, errors.Wrap(err, "failed to parse job config template text")
-	}
-	cl.jobConfigTemplate = jct
-	ict, err := template.New("irods_config").Parse(IRODSConfigTemplateText)
-	if err != nil {
-		return cl, errors.Wrap(err, "failed to parse irods config template text")
-	}
-	cl.irodsConfigTemplate = ict
-	return cl, err
-}
-
-// GenerateCondorSubmit returns a string (or error) containing the contents
-// of what should go into an HTCondor submission file.
-func (cl *CondorLauncher) GenerateCondorSubmit(submission *model.Job) (*bytes.Buffer, error) {
-	var buffer bytes.Buffer
-	err := cl.submissionTemplate.Execute(&buffer, submission)
-	if err != nil {
-		return &buffer, errors.Wrap(err, "failed to apply data to the submission template")
-	}
-	return &buffer, nil
-}
-
-type scriptable struct {
-	model.Job
-	DC []model.VolumesFrom
-	CI []model.ContainerImage
-}
-
-// GenerateJobConfig creates a string containing the config that gets passed
-// into the job.
-func (cl *CondorLauncher) GenerateJobConfig() (*bytes.Buffer, error) {
-	var buffer bytes.Buffer
-	err := cl.jobConfigTemplate.Execute(&buffer, cl.cfg)
-	if err != nil {
-		return &buffer, errors.Wrap(err, "failed to apply data to the job config template")
-	}
-	return &buffer, nil
-}
-
-type irodsconfig struct {
-	IRODSHost string
-	IRODSPort string
-	IRODSUser string
-	IRODSPass string
-	IRODSZone string
-	IRODSBase string
-	IRODSResc string
-}
-
-// GenerateIRODSConfig returns the contents of the irods-config file as a string.
-func (cl *CondorLauncher) GenerateIRODSConfig() (*bytes.Buffer, error) {
-	c := &irodsconfig{
-		IRODSHost: cl.cfg.GetString("irods.host"),
-		IRODSPort: cl.cfg.GetString("irods.port"),
-		IRODSUser: cl.cfg.GetString("irods.user"),
-		IRODSPass: cl.cfg.GetString("irods.pass"),
-		IRODSBase: cl.cfg.GetString("irods.base"),
-		IRODSResc: cl.cfg.GetString("irods.resc"),
-		IRODSZone: cl.cfg.GetString("irods.zone"),
-	}
-	var buffer bytes.Buffer
-	err := cl.irodsConfigTemplate.Execute(&buffer, c)
-	if err != nil {
-		return &buffer, errors.Wrap(err, "failed to apply data to the irods config template")
-	}
-	return &buffer, err
-}
-
-// CreateSubmissionDirectory creates a directory for a submission and returns the path to it as a string.
-func (cl *CondorLauncher) CreateSubmissionDirectory(s *model.Job) (string, error) {
-	dirPath := s.CondorLogDirectory()
-	if path.Base(dirPath) != "logs" {
-		dirPath = path.Join(dirPath, "logs")
-	}
-	err := os.MkdirAll(dirPath, 0755)
-	if err != nil {
-		return "", errors.Wrapf(err, "failed to create the directory %s", dirPath)
-	}
-	return dirPath, err
 }
 
 // CreateSubmissionFiles creates the iplant.cmd file inside the
 // directory designated by 'dir'. The return values are the path to the iplant.cmd
 // file, and any errors, in that order.
 func (cl *CondorLauncher) CreateSubmissionFiles(dir string, s *model.Job) (string, string, string, error) {
-	cmdContents, err := cl.GenerateCondorSubmit(s)
+	cmdContents, err := GenerateCondorSubmit(SubmissionTemplate, s)
 	if err != nil {
 		return "", "", "", err
 	}
 
-	jobConfigContents, err := cl.GenerateJobConfig()
+	jobConfigContents, err := GenerateJobConfig(JobConfigTemplate, cl.cfg)
 	if err != nil {
 		return "", "", "", err
 	}
@@ -274,7 +128,7 @@ func (cl *CondorLauncher) CreateSubmissionFiles(dir string, s *model.Job) (strin
 		return "", "", "", err
 	}
 
-	irodsContents, err := cl.GenerateIRODSConfig()
+	irodsContents, err := GenerateIRODSConfig(IRODSConfigTemplate, cl.cfg)
 	if err != nil {
 		return "", "", "", err
 	}
@@ -331,7 +185,7 @@ func (cl *CondorLauncher) submit(cmdPath string, s *model.Job) (string, error) {
 }
 
 func (cl *CondorLauncher) launch(s *model.Job) (string, error) {
-	sdir, err := cl.CreateSubmissionDirectory(s)
+	sdir, err := CreateSubmissionDirectory(s)
 	if err != nil {
 		log.Errorf("%+v\n", errors.Wrap(err, "failed to create submission directory"))
 		return "", err
@@ -487,6 +341,7 @@ func main() {
 	var (
 		cfgPath     = flag.String("config", "", "Path to the config file. Required.")
 		showVersion = flag.Bool("version", false, "Print the version information")
+		err         error
 	)
 
 	flag.Parse()
@@ -521,10 +376,7 @@ func main() {
 	defer client.Close()
 
 	localfs := &osys{}
-	launcher, err := New(cfg, client, localfs)
-	if err != nil {
-		log.Fatalf("%+v\n", err)
-	}
+	launcher := New(cfg, client, localfs)
 
 	launcher.client.SetupPublishing(exchangeName)
