+++ stops_test.go
@@ -461,10 +461,7 @@ func TestCondorID(t *testing.T) {
 func TestExecCondorQ(t *testing.T) {
 	inittests(t)
 	filesystem := newtsys()
-	cl, err := New(cfg, nil, filesystem)
-	if err != nil {
-		t.Error(err)
-	}
+	cl := New(cfg, nil, filesystem)
 	output, err := cl.ExecCondorQ()
 	if err != nil {
 		t.Error(err)
@@ -512,10 +509,7 @@ func TestExecCondorQ(t *testing.T) {
 func TestExecCondorRm(t *testing.T) {
 	inittests(t)
 	filesystem := newtsys()
-	cl, err := New(cfg, nil, filesystem)
-	if err != nil {
-		t.Error(err)
-	}
+	cl := New(cfg, nil, filesystem)
 	actual, err := cl.ExecCondorRm("foo")
 	if err != nil {
 		t.Error(err)
@@ -537,10 +531,7 @@ func TestStopHandler(t *testing.T) {
 	}
 	inittests(t)
 	filesystem := newtsys()
-	cl, err := New(cfg, nil, filesystem)
-	if err != nil {
-		t.Error(err)
-	}
+	cl := New(cfg, nil, filesystem)
 	stopMsg := messaging.StopRequest{
 		InvocationID: "b788569f-6948-4586-b5bd-5ea096986331",
 	}
